+++ xdef/changelog.md
@@ -1,7 +1,8 @@
 `# Version ${version}, release-date ${release.date}
 
 # Version 42.2.1, release-date 2024-07-03
-* .
+* Fixed error in compilation of type declaration with expression containing
+  `CHECK` operator.
 
 # Version 42.2.0, release-date 2024-06-27
 * Fixed a bug that the `onXMLError` section in the X-script was not compiled and